+++ ABP/abp.h
@@ -22,7 +22,7 @@ pNodoA* InsereArvore(pNodoA *a, tipoinfo ch);
 void preFixado(pNodoA *a);
 void posFixado(pNodoA *a);
 void Central(pNodoA *a);
-pNodoA* consultaABP(pNodoA *a, tipoinfo chave, unsigned int*comp);
+pNodoA* consultaABP(pNodoA *a, tipoinfo chave, unsigned long int*comp);
 pNodoA* consultaABP2(pNodoA *a, tipoinfo chave);
 
 #endif
